Move CoinVerifierApiClient
